+++ sdk/search/Azure.Search.Documents/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Added
 
+- Add `SearchIndexingBufferedSender<T>` to make indexing lots of documents fast and easy.
 - Add support to `FieldBuilder` to define search fields for `Microsoft.Spatial` types without an explicit assembly dependency.
 - Add support to `SearchFilter` to encode geometric types from `Microsoft.Spatial` without an explicit assembly dependency.
 - Add `IndexingParameters.IndexingParametersConfiguration` property to define well-known properties supported by Azure Cognitive Search.